Begin supporting type "notations".
The rough idea is that this allows constraints to be specified inline in
types rather than in the initial qualifiers.  For example, a identity
function for types in  both Bounded and Show could be typed as

  t {Show,Bounded} -> t

By itself, this is probably not much use; I expect it to pay off with
the (coming) linearity constraints however.