Support functional notation without functional dependencies.
This allows any class to be used in place of a type.  Note that repeated
instances of the same class do not denote the same type; so, for
example, the type

   Num -> Num

is interpreted as the type

   (Num t, Num u) => t -> u

This may seem of limited value on its own, but should prove essential
for linearity.