Fix treatment of dropped type variables in type inference.
For the canonical example, consider:

    x = null []

Here we know that the empty list [] must be of some type [a], and that
null must be of type [a] -> Bool, but have no way to fix a particular
type [a].  This leads to problems in specialization: we end up with an
unbound type variable appearing in the XMPEG expression.

This bug also appears in the treatment of pattern bindings: any
polymorphic pattern bindings would result in unbound type variables in
specialization.

To avoid this, we now track each type variable that "disappears" in an
application, see which of them are actually unconstrained and
unquantified at generalization, and replace those types with 'Zero', the
empty type.  Since those type variables should actually have no
consequence for the semantics of terms, this choice shouldn't matter.

The implementation is more than a bit awkward.