Replace unbound type variables with Zero after type inference.
We can have unbound variables after type inference in cases like the
following:

    (f, g) = (\x -> x, \y -> y)

In this case, we infer that the pair (call it 'p') has type '(t -> t, u
-> u)'; we bind 'f' to 'fst p' and 'g' to 'snd p'.  However, in those
bindings, we have no information to fix the types of 'u' and 't',
respectively.  Parametricity tells us that any type will do, and so we
(not entirely arbitrarily) pick the empty type Zero.