Begin reflecting MPEG normalization in XMPEG syntax.
The core change here is that we require that bindings be of the form

    P <- x

rather than

    P <- E

in XMPEG.  This simplifies cases in post-typechecking stages, and
reflects a normalization step that was (almost) already performed.

The only remaining use of bindings

   P <- E

in the IMPEG stages of the compiler is to limit generalization.
Concretely, suppose we have a source language term

    case maybeIx 0 of
      Nothing -> ...
      Just i -> ... i ...

If we attempt to rewrite this using let, we get something like the
following:

    {let scrutinee = maybeIx 0 =>
     (Nothing <- scrutinee => ...   |
      Just  i <- scrutinee => .. i ..)}

But now, generalization will become obnoxious: `scrutinee` will be
generalized, with an `Index` predicate, and the `Nothing` branch won't
fix that variable, leaving an ambiguous type and unsolved predicate in
the resulting term.