Fix treatment of qualifiers on constructors; begin fixing treatment of primitive constructors.
The treatment of constructors with qualified types was completely
broken.  It's now fixed, but there's no longer any attempt to tell a
story about existential types.

The treatment of primitive constructors is now somewhat unbroken up to
specialization, but still broken in specialization and after.  The
current XMPEG AST doesn't capture primitive constructors at all, which
seems like a mistake to me.