Temporary fix for recursive datatypes.
TopDecl:149 The problem is that we need to return the correct type structure.
kids params, vs, ts are the 4 things used in that function

params may not contain the entire structural information hence we now use ts
however we have to select a (unique) identifier name for each element of ts that we can stuff in
the sharing [[Id]].

tytoi is the function that is expected to return (unique) identifier.
ts is a [Located Type]
   for TyVar its easy
   for TyApp it gets tricky as there can be a deep nesting.