Allow multiple primitive types in a single declaration.
This also changes the syntax of primitive type declarations slightly.
Previously, we had

    primitive type <tyexp>

where <tyexp> could be things like "Foo :: * -> *" or "Bar (n :: nat)".
Now, the syntax is

    primitive type tycon1, tycon2, .. :: kind

requiring the kind to be written out, and without introducing random
names for parameters or other mysteries of the tyexp grammar.

Along the way: clean a small amount of cruft out of the prelude; restore
BE and LE as primitive types; update primitive type declarations to
match new primitive type declaration form.