Support labeled fields for ordinary data types.
This patch adds support for labeled fields, vaguely a la Haskell, to
ordinary data types.

When declaring a data type, you can optionally give field labels.  For
example:

    data Cake = Birthday[ age :: Unsigned | candles :: Bool]
              | Chocolate

Labeled arguments are also positional, so both of the following are
valid Cake terms:

    Birthday[age = 12 | candles = True]
    Birthday 64 False

Similarly, in pattern matching, both positional and labeled patterns are
supported; missing fields in labeled patterns are treated as wildcards.

Instances of the `Select` and `Update` class are generated
automatically.  So, you could write:

    nextYear'sCake cake = cake[age = cage.age + 1]

The generated `select` and `update` methods will fail with pattern match
errors when applied to values that lack the corresponding fields.  The
following code would fail at run-time:

    nextYear'sCake Chocolate

This is in contrast to the situation with bitdata, but in line with the
approach in Haskell.  I am deferring further attempts to get it right to
the Grand Unification of data types.