Replace the prelude definition of fromBits for Bool.
The old (i.e., now replaced) version was as follows:

fromBits B0 = False
fromBits B1 = True

Matching on literals is implemented via equality tests,
so this translates as roughly equivalent to:

fromBits x = case x==B0 of
               True -> False
               _    -> case x==B1 of
                         True -> True
                         _    -> error "unreachable"

We know that, for any Bit 1 value,  one of those two
equality tests is going to succeed.  Or, at least, we
assume that the implementation of == will do what we
expect and so we infer that the error in the last line
will never occur in practice.  But our compiler can't
guarantee that, so it keeps the code for error case in
the compiled binary, just to be sure.  The resulting
MIL program will end up including a call to the "loop"
primitive, and although we don't expect that call will
ever be made, we will get a linker error if we don't
provide an implementation for loop.

The new version of fromBits is as follows:

fromBits b = if b==B0 then False else True

This performs a single equality test and does not leave
room for an error case.  As a result, programs that
use this version can be compiled without requiring
an implementation of loop.