Treat LC as a strange output format for XMPEG.
In the beginning there was MPEG.  MPEG had many variations---implicitly
polymorphic, explicitly polymorphic, monomorphic---but they were all
aspects of Habit's one intermediate language.

Then, there was lambda case.  Lambda case was like MPEG, but broken; for
lo, lambda case included all of MPEG:

     [[ ^e ]]          = e
     [[ fail ]]        = case False of True -> abort
     [[ p <- w => m ]] = case w of p -> [[ m ]]
     [[ let ds => m ]] = let [[ ds ]] in [[ m ]]
     [[ m | n ]]       = [[ m ]] | [[ n ]]

But whereas in MPEG, the extent of failure was limited, in lambda case
it is not, so all matches should really be translated

     [[ { m } ]]       = ([[ m ]] | abort)

But indeed they never were.

Then there was LC, which is like lambda case but is printed slightly
differently, so it like MPEG but broken but printed slightly
differently.

And adding features to the compiler was made incrementally more and more
annoying, as there were more and more syntax trees, and more and more
validation steps, but no more and no more information about the language
along the way.

And so finally the compiler authors spake: enough!  And lambda case was
no longer part of the front end, but rather a funny way of printing
MPEG.  And many passes were needed no longer, and many lines of code
were removed.  And it was good.

Or at least incrementally better.