First attempt at support for partial data type declarations.
The new feature here is that data type declarations can include contexts
which (a) must be true when a value of the data type is constructed,
and (b) can assumed to be true when such a value is used.  The
canonical (if controversial) example is binary search trees:

    data Ord a => BST a = Leaf | Node a (BST a) (BST a)

This declaration introduces the `BST a` type, but only for types `a`
that are ordered.  This means that, for example, the insert function has
type:

    insert :: a -> BST a -> BST a

No `Ord a` constraint in sight, because it's implied by the type `BST
a`.  See examples/bst.hb for all the details; in particular, there's a
`Functor` instance which depends on the implicit constraints.

  My implementation is based on the two-parameter version of
inhabitation constraints.  So, the above type is implicitly interpreted
as:

    insert :: BST@a => a -> BST a -> BST a

while the type of `fmap` is interpreted as

    fmap :: (f@a, f@b) => (a -> b) -> f a -> f b

Instances of @ are generated automatically.  For an total data type
these are trivial.  For example,

    data List a = Nil | Cons a (List a)

induces

    instance List @ a

For partial data types, these are somewhat more interesting.  The
declaration of BSTs above generates:

    instance BST @ a if Ord a
    else BST @ a fails

Primitive types are assumed to be total.  This should be fixed in the
future, but the current syntax leaves no room for a context!

As we might have expected from Mark's experiments, other than
introducing the (@) class there are no changes to existing
examples (including all the prelude types and the gdtac example, which
does quite a bit of type-level stuff).  However, there does seem to be
some grit in the gears when it comes to interacting with the back-end,
so some further patches may be in order.