Simplication for GFP axioms excludes requirements of the conclusion.
  (From src/Solver/Validation.hs:)

This next bit is a hack, if a necessary one (at the moment).  The
problem arises when we have matching pairs of requirement and instance,
such as in the current approach to partial data types:

    @ BST a requires Ord a.
    @ BST a if Ord a; @ BST a fails!

Now, consider the simplification of the instance: because this is a
derived instance, it gets a GFP interpretation.  But (by the
requirement) `@ BST a requires Ord a`, and the `Ord a` hypothesis can be
eliminated.  This _would_ give us the instance `@ BST a; @ BST a fails.`
_except that_ now the instance doesn't satisfy the requirement, and so
it's rejected.

The hacky solution is that, only when considering the GFP of an
instance, we eliminate all requirements that match the conclusion... in
this example, we drop the `@ BST a requires Ord a` requirement, because
our instance concludes `@ BST a`.  Not really sure about the semantic
consequences.. this may mean that GFP instances aren't really GFP
instances, but I can't think of an argument pro or con at the moment.