Generate fewer cyclic proofs.
The initial source of pain here is that improvement may end up creating
duplicate assumptions in the trail.  For example, if we start out with
two goals

    F a b, F a Bool

along with the typical functional dependency, we will eventually learn
that b +-> Bool.  Now, we have a danger of creating a cyclic proof, in
which we prove goal 1 in terms of goal 2 and vice versa.

The new twist is that assumptions may themselves introduce chains of
other assumptions from requirements.  So even if we avoid the cycle
above, we could have gotten a cycle indirectly via a requirement.

The solution I have is somewhat ad-hoc: at the moment, only nodes in the
forest, not nodes on the trail, track which assumptions they introduce.
So, when we return to a forest node, we check to see whether any of the
assumptions introduced by that node have since been invalidated.  If
they have, we invalidate every other assumption introduced by the same
node.

I'm not convinced this is really the right behavior, but I think it may
be about the best I can do with the current setup.  The real point, I
think, is that this information should *all* be in the trail, while the
forest should basically be an artifact of constructing the trail.