Generate more improvements (without more cyclic proofs).
This commit revisits the problem first attacked in d9931d.  As there,
the difficulty is cyclic proofs caused by requirement templates built
from subsequently invalidated assumptions.  The solution in that commit
seemed draconian to me at the time, and turns out to have resulted in
discarding valid improvements along with invalid assumptions.  This
patch replaces that mechanism with filtering the results of requirement
templates looking for invalidated assumptions when those results are
used.

New tests: cyclic1 demonstrates the original problem, impr10
demonstrates the improvements that were being lost with the original
solution.